rpaframework-actions/requirements.txt
pyinstaller==6.3.0
rpaframework==28.2.0
